+++ lcgrand.h
@@ -3,6 +3,6 @@
 #define METALLER2_LCGRAND_H
 
 extern long zrng[];  // Declare zrng array
-double LCGrand(int num);  // Declare LCGrand function
+float LCGrand(int num);  // Declare LCGrand function
 
 #endif //METALLER2_LCGRAND_H
